Allow specifying of the Page model through the config in order to make scout searchable optional
