+++ src-php/Config/novapages.php
@@ -1,6 +1,10 @@
 <?php
 
 return [
+    'models' => [
+        'page' => 'Dewsign\NovaPages\Models\Page',
+        // 'page' => 'Dewsign\NovaPages\Models\PageSearchable',
+    ],
     'repeaters' => [],
     'replaceRepeaters' => false,
     'homepageSlug' => 'homepage',
