+++ src-php/Database/seeds/PageSeeder.php
@@ -17,13 +17,15 @@ class PageSeeder extends Seeder
      */
     public function run()
     {
-        factory(Page::class, 100)->create()->each(function ($page) {
+        $pageClass = config('novapages.models.page', Page::class);
+
+        factory($pageClass, 100)->create()->each(function ($page) {
             $page->repeaters()->saveMany(factory(Repeater::class, rand(0, 5))->create()->each(function ($repeater) {
                 $repeater->type()->associate(factory(AvailableBlocks::random())->create())->save();
             }));
         });
 
-        Page::inRandomOrder()->take(rand(25, 75))->get()->each(function ($page) {
+        app($pageClass)::inRandomOrder()->take(rand(25, 75))->get()->each(function ($page) {
             $page->parent()->associate(Page::inRandomOrder()->where('id', '<>', $page->id)->first());
             $page->save();
         });
