+++ src-php/Models/Page.php
@@ -13,13 +13,11 @@ use Maxfactor\Support\Model\Traits\WithPrioritisation;
 use Maxfactor\Support\Webpage\Traits\HasMetaAttributes;
 use Maxfactor\Support\Webpage\Traits\MustHaveCanonical;
 use Dewsign\NovaRepeaterBlocks\Traits\HasRepeaterBlocks;
-use Dewsign\NovaPages\IndexConfigurators\PageIndexConfigurator;
 
 class Page extends Model
 {
     use HasSlug;
     use HasParent;
-    use Searchable;
     use CanBeFeatured;
     use HasActiveState;
     use HasMetaAttributes;
@@ -27,45 +25,7 @@ class Page extends Model
     use MustHaveCanonical;
     use WithPrioritisation;
 
-    protected $indexConfigurator = PageIndexConfigurator::class;
-
-    // Mapping for a model fields.
-    protected $mapping = [
-        'properties' => [
-            'text' => [
-                'type' => 'text',
-                'fields' => [
-                    'raw' => [
-                        'type' => 'keyword',
-                    ]
-                ]
-            ],
-        ]
-    ];
-
-    /**
-     * Get the indexable data array for the model.
-     *
-     * @return array
-     */
-    public function toSearchableArray()
-    {
-        $this->repeaters;
-
-        $searchable = $this->toArray();
-
-        $searchable = array_except($searchable, [
-            'active',
-            'browser_title',
-            'h1',
-            'meta_description',
-            'nav_title',
-            'canonical',
-            'parent',
-        ]);
-
-        return $searchable;
-    }
+    protected $table = 'pages';
 
     /**
      * The attributes that are mass assignable.
@@ -132,7 +92,7 @@ class Page extends Model
      * Recursively add parent pages to the breadcrumb seed
      *
      * @param Illuminate\Support\Collection $seed
-     * @param Dewsign\NovaPages\Models\Page $item
+     * @param Dewsign\NovaPages\Facades\Page $item
      * @return Illuminate\Support\Collection
      */
     private function seedParent(&$seed, $item)
