+++ src-php/Tests/PageTest.php
@@ -2,7 +2,7 @@
 
 namespace Dewsign\NovaPages\Tests;
 
-use Dewsign\NovaPages\Models\Page;
+use Dewsign\NovaPages\Facades\Page;
 use Illuminate\Foundation\Testing\TestCase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
